+++ src/main/java/com/exmple/kyudo_app/domain/model/TeamMember.java
@@ -1,4 +1,4 @@
-package com.exmple.kyudo_app.model;
+package com.exmple.kyudo_app.domain.model;
 
 import jakarta.persistence.*;
 import lombok.Getter;
